[CHORE] folder structure edited
