+++ Assingment_Tving/Assingment_Tving/Screen/View/Main/Kids/KidsView.swift
@@ -1,5 +1,5 @@
 //
-//  MainView.swift
+//  KidsView.swift
 //  Assingment_Tving
 //
 //  Created by 이성민 on 2023/05/03.
