+++ Assingment_Tving/Assingment_Tving/Screen/View/Main/Paramount/ParamountView.swift
@@ -1,5 +1,5 @@
 //
-//  MainView.swift
+//  ParamountView.swift
 //  Assingment_Tving
 //
 //  Created by 이성민 on 2023/05/03.
