+++ Assingment_Tving/Assingment_Tving/Screen/ViewController/Main/LiveViewController.swift
@@ -1,5 +1,5 @@
 //
-//  MainView.swift
+//  LiveViewController.swift
 //  Assingment_Tving
 //
 //  Created by 이성민 on 2023/05/03.
