+++ Assingment_Tving/Assingment_Tving/Screen/ViewController/Main/TVViewController.swift
@@ -1,5 +1,5 @@
 //
-//  MainView.swift
+//  TVViewController.swift
 //  Assingment_Tving
 //
 //  Created by 이성민 on 2023/05/03.
